create simple block page with route block/:height and dynamically fetch and render data - no styling
