--- client/src/components/Block.vue
@@ -1,17 +1,0 @@
-<template>
-  <h1>This is the {{ $route.params.height }} BLOCK page</h1>
-</template>
-
-<script>
-export default {
-  data() {
-    return {
-      
-    }
-  }
-}
-</script>
-
-<style>
-
-</style>
