+++ client/src/main.js
@@ -2,7 +2,7 @@ import Vue from "vue";
 import VueRouter from 'vue-router';
 import App from "./App.vue";
 import Home from "./components/Home.vue";
-import Block from "./components/Block.vue";
+import BlockPage from "./components/BlockPage/BlockPage.vue";
 
 
 //set up VueRouter
@@ -10,7 +10,7 @@ Vue.use(VueRouter);
 
 const routes = [
   { path: '/', component: Home},
-  { path: '/block/:height', component: Block }
+  { path: '/blocks/:height', component: BlockPage },
 ]
 
 const router = new VueRouter({
